requirements.txt
requests>=2.31
python-dotenv>=1.0
Pillow>=10.2
Unidecode>=1.3
